+++ pkg/commands/ngwaf/workspace/workspace_test.go
@@ -12,7 +12,7 @@ import (
 	sub "github.com/fastly/cli/pkg/commands/ngwaf/workspace"
 	fstfmt "github.com/fastly/cli/pkg/fmt"
 	"github.com/fastly/cli/pkg/testutil"
-	"github.com/fastly/go-fastly/v15/fastly/ngwaf/v1/workspaces"
+	"github.com/fastly/go-fastly/v16/fastly/ngwaf/v1/workspaces"
 )
 
 const (