+++ pkg/commands/stats/aggregate_test.go
@@ -7,7 +7,7 @@ import (
 	"io"
 	"testing"
 
-	"github.com/fastly/go-fastly/v15/fastly"
+	"github.com/fastly/go-fastly/v16/fastly"
 
 	"github.com/fastly/cli/pkg/app"
 	"github.com/fastly/cli/pkg/global"